pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.8.0,<3.9"
numpy = "^1.22.2"
replit = "^3.2.4"
urllib3 = "^1.26.12"
requests = "^2.28.1"
python-dotenv = "^0.21.0"
PyGithub = "^1.57"
PyYAML = "^6.0"
pandas = "^1.5.2"
scikit-learn = "^1.2.0"
sklearn = "^0.0.post1"
fancyimpute = "^0.7.0"
Edward = "^1.3.5"

[tool.poetry.dev-dependencies]
debugpy = "^1.6.2"
replit-python-lsp-server = {extras = ["yapf", "rope", "pyflakes"], version = "^1.5.9"}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"